+++ src/game/components.hpp
@@ -10,7 +10,7 @@ struct PortalTag{};
 struct PortalTag2{};
 
 struct PlayerTag{
-    int dmg = 1;
+    float dmg = 1;
     float attackSpeed = 0.1f;
     float projectileCooldown = 0.0f;
     float radius = 5.0f;
